services/license-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.9.0
pydantic-settings==2.5.0
redis==5.0.0
aioredis==2.0.1
cryptography==41.0.7
python-json-logger==2.0.7
pyjwt==2.8.0
